+++ signal-protocol-client/src/identity/state.ts
@@ -1,6 +1,6 @@
-import { SignalProtocolStore } from '@app/identity/signal-store'
+import { SignalProtocolStore } from '@app/signal/signal-store'
 import { BehaviorSubject } from 'rxjs'
-import { SignalDirectory } from './signal-directory'
+import { SignalDirectory } from '../signal/signal-directory'
 
 export const directorySubject = new BehaviorSubject<SignalDirectory | null>(null)
 export const signalStore = new SignalProtocolStore()
