+++ signal-protocol-client/src/signal/signal-directory.ts
